Sincronizacion offline-first de tickets
El front pasa a poder vender con internet caido: arma el ticket en el
dispositivo, lo guarda en su cola y lo manda en lote a
POST /api/ventas/v1/sync cuando vuelve la conexion. Eso obliga a romper
cuatro reglas del resto del sistema, cada una a proposito y con su
motivo escrito en el codigo y en ARCHITECTURE.md.

El id lo genera el cliente, y es UUIDv7. El ticket ya existia en el
dispositivo antes de que el backend supiera de el. Que sea v7 no es
cosmetico: la PK es el clustered index de InnoDB y un v4 es aleatorio de
punta a punta, asi que cada insercion cae en una pagina al azar. Los 48
bits altos de un v7 son el timestamp. Hibernate 7 trae VERSION_7 pero no
sirve: un generador genera siempre y pisaria el uuid del ticket, asi que
el id es asignado y existe shared/Uuid7. Como consecuencia, las tres
entidades que reciben ids del front implementan Persistable; sin eso
Spring Data decide que una entidad con id no es nueva y cada save() pasa
por merge(), o sea un SELECT que no devuelve nada antes de cada INSERT.
Un ticket de diez lineas son once consultas de mas.

El vendedor viaja en el payload. Es la excepcion a "la identidad sale
siempre del JWT", y la unica del sistema. Con una terminal y cambio de
turno, el que esta logueado cuando vuelve internet no es el que hizo los
tickets encolados. Se valida que exista y nada mas: exigir que este
activo dejaria los tickets de un empleado dado de baja rechazados para
siempre, y la mercaderia salio igual. La mitigacion es auditoria, no
permiso: id_usuario_sync guarda quien sincronizo y origen marca por
donde entro.

Las fechas las manda el dispositivo. created_at pasa a significar cuando
ocurrio el ticket y deleted_at cuando se anulo; el dato nuevo es
sincronizado_en. Un lote son veinte tickets que pasaron en momentos
distintos: si la fecha la pusiera el INSERT, los veinte quedarian
fechados en el mismo segundo y dos dias despues de lo que ocurrio. La
fecha baja por toda la cadena —el movimiento de stock y el de caja
tambien—, porque si no el kardex muestra la mercaderia saliendo despues
de la venta que la saco. Como la fecha es ahora un dato del cliente, se
valida: desfase futuro, antiguedad maxima, y una anulacion nunca
anterior a su ticket.

Y un faltante de stock no rechaza el ticket. Esto es una caja
registradora, no un ecommerce: la venta es fisica y ya ocurrio. Si el
sistema decia 4 y se vendieron 10, el sistema estaba mal y las 10
estaban en la gondola. Entra un AJUSTE por la diferencia y despues la
VENTA completa, en ese orden. Registrar la venta parcial estaria mal por
dos motivos: el kardex dejaria de sumar al stock, y la anulacion
repondria 4 en vez de 10, porque la reversa trabaja sobre los
movimientos de tipo VENTA.

Ademas:

Una transaccion por evento, no una por lote. Un lote a medio procesar
tiene que dejar la base consistente, porque es exactamente lo que pasa
cuando la respuesta se pierde en el camino. Por eso ProcesadorEventoSync
es un bean aparte: una llamada interna no pasa por el proxy de Spring y
no abriria ninguna transaccion nueva.

Idempotencia por tipo + uuid, con la misma respuesta que la primera vez.
Un OK pelado en el reenvio haria que el unico caso donde la idempotencia
importa —la respuesta original se perdio— fuera el unico donde el front
nunca se entera de que el ticket quedo marcado.

Los eventos se aplican en el orden en que pasaron en el local, no en el
del array: es lo que hace que un ticket y su anulacion se apliquen bien,
y que la apertura de un turno llegue antes que sus tickets.

Anular deja de ser exclusivo de ADMIN: un EMPLEADO puede anular las
propias dentro de 7 dias medidos sobre la apertura de su turno. La regla
vale igual por el panel que por el sync, o anular offline seria la forma
de saltear el permiso. Una venta cobrada ahora se anula, y si movio
efectivo la plata vuelve por el turno abierto de hoy: el corte viejo ya
esta firmado.

Los importes de ventas pasan a DECIMAL. El total lo suma tambien el
front y las dos cuentas tienen que dar exactamente igual; con float cada
diferencia de redondeo apareceria como un ticket "para revisar" que no
tiene nada. El resto del sistema sigue en float y las fronteras pasan
por shared/Importes, que evita la trampa de BigDecimal.valueOf(float).

El descuento de stock y la anulacion se extrajeron a colaboradores
propios: hay dos caminos que venden y tres que anulan, y son la parte
que toma locks de fila. Dos copias de esos loops es la forma mas
probable de que un dia tomen los locks en ordenes distintos y la base
empiece a matar transacciones por deadlock.

Migraciones 12 y 13, verificadas contra una base real: el esquema que
producen es identico byte a byte al que arma 00_init_limpio.sql desde
cero. La 13 avisa de algo contraintuitivo: cabecera y lineas se
redondean por separado, asi que la conversion puede descuadrar una venta
que en float cuadraba.

88 tests nuevos, y cada fase verificada contra MySQL ademas de con
mocks. Varios de los problemas de arriba —el SELECT antes de cada
INSERT, la fecha del retiro del corte— solo se vieron corriendo contra
la base de verdad.

+++ src/main/java/com/SolucionesInformaticasBA/minimarket/modules/ventas/api/dto/CobrarVentaRequest.java
@@ -1,5 +1,7 @@
 package com.SolucionesInformaticasBA.minimarket.modules.ventas.api.dto;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
 import lombok.Data;
@@ -13,7 +15,7 @@ public class CobrarVentaRequest {
      * como si fuera un dato real. Con esos medios se ignora.
      */
     @PositiveOrZero
-    private Float montoRecibido;
+    private BigDecimal montoRecibido;
 
     @NotBlank
     private String metodoPago;
