+++ src/main/java/com/SolucionesInformaticasBA/minimarket/modules/ventas/api/dto/CobrarVentaResponse.java
@@ -1,5 +1,7 @@
 package com.SolucionesInformaticasBA.minimarket.modules.ventas.api.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 import lombok.Data;
 
@@ -7,5 +9,5 @@ import lombok.Data;
 @Builder
 public class CobrarVentaResponse {
     private VentaResponse venta;
-    private float cambio;
+    private BigDecimal cambio;
 }
